+++ crates/polars-plan/src/logical_plan/mod.rs
@@ -193,7 +193,6 @@ pub enum DslPlan {
     HStack {
         input: Arc<DslPlan>,
         exprs: Vec<Expr>,
-        schema: SchemaRef,
         options: ProjectionOptions,
     },
     /// Remove duplicates from the table
